hpcgap: merge src/stats.c
